requirements.txt
selenium
lxml
beautifulsoup4
Pillow
instabot
pyyaml
